+++ Ejercicio del Mes/Form1.cs
@@ -42,8 +42,9 @@ namespace Ejercicio_del_Mes
 
         private void btn_calcular_Click(object sender, EventArgs e)
         {
-            float horasEstancia = float.Parse(txt_horas.Text);
-            ClaseVehiculo claseVehiculo = null;
+            int horasEstancia = int.Parse(txt_horas.Text);
+            ClaseVehiculo claseVehiculo = null; 
+
             //Aqui trabajara Kevin
             if (cboVehiculo.Text == "Motocicleta")
             {
